+++ src/screenComponents/mainScreenControls.cpp
@@ -11,7 +11,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     setPosition(-20, 70, ATopRight);
 
     // Set which buttons appear when opening the main screen controls.
-    open_button = new GuiToggleButton(this, "MAIN_SCREEN_CONTROLS_SHOW", "Main screen", [this](bool value)
+    open_button = new GuiToggleButton(this, "MAIN_SCREEN_CONTROLS_SHOW", tr("MainScreenControls", "Main screen"), [this](bool value)
     {
         for(GuiButton* button : buttons)
             button->setVisible(value);
@@ -32,7 +32,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     open_button->setSize(GuiElement::GuiSizeMax, 50);
 
     // Front, back, left, and right view buttons.
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_FRONT_BUTTON", "Front", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_FRONT_BUTTON", tr("MainScreenControls", "Front"), [this]()
     {
         if (my_spaceship)
         {
@@ -40,7 +40,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
         }
         closePopup();
     }));
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_BACK_BUTTON", "Back", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_BACK_BUTTON", tr("MainScreenControls", "Back"), [this]()
     {
         if (my_spaceship)
         {
@@ -48,7 +48,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
         }
         closePopup();
     }));
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_LEFT_BUTTON", "Left", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_LEFT_BUTTON", tr("MainScreenControls", "Left"), [this]()
     {
         if (my_spaceship)
         {
@@ -56,7 +56,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
         }
         closePopup();
     }));
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_RIGHT_BUTTON", "Right", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_RIGHT_BUTTON", tr("MainScreenControls", "Right"), [this]()
     {
         if (my_spaceship)
         {
@@ -69,7 +69,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     // option in the main screen controls.
     if (my_player_info->crew_position[weaponsOfficer] || my_player_info->crew_position[tacticalOfficer] || my_player_info->crew_position[singlePilot])
     {
-        buttons.push_back(new GuiButton(this, "MAIN_SCREEN_TARGET_BUTTON", "Target lock", [this]()
+        buttons.push_back(new GuiButton(this, "MAIN_SCREEN_TARGET_BUTTON", tr("MainScreenControls", "Target lock"), [this]()
         {
             if (my_spaceship)
             {
@@ -81,7 +81,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     }
 
     // Tactical radar button.
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_TACTICAL_BUTTON", "Tactical", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_TACTICAL_BUTTON", tr("MainScreenControls", "Tactical"), [this]()
     {
         if (my_spaceship)
         {
@@ -92,7 +92,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     tactical_button = buttons.back();
 
     // Long-range radar button.
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_LONG_RANGE_BUTTON", "Long Range", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_LONG_RANGE_BUTTON", tr("MainScreenControls", "Long Range"), [this]()
     {
         if (my_spaceship)
         {
@@ -103,7 +103,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     long_range_button = buttons.back();
 
     // Far-range radar button.
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_FAR_RANGE_BUTTON", "Far Range", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_FAR_RANGE_BUTTON", tr("MainScreenControls", "Far Range"), [this]()
     {
         if (my_spaceship)
         {
@@ -114,7 +114,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     far_range_button = buttons.back();
 
     // Far-range radar button.
-    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_TARGET_ANALYSIS_BUTTON", "Target analysis", [this]()
+    buttons.push_back(new GuiButton(this, "MAIN_SCREEN_TARGET_ANALYSIS_BUTTON", tr("MainScreenControls", "Target analysis"), [this]()
     {
         if (my_spaceship)
         {
@@ -128,7 +128,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
     // on the main screen.
     if (my_player_info->crew_position[relayOfficer] || my_player_info->crew_position[operationsOfficer] || my_player_info->crew_position[singlePilot])
     {
-        buttons.push_back(new GuiButton(this, "MAIN_SCREEN_SHOW_COMMS_BUTTON", "Show comms", [this]()
+        buttons.push_back(new GuiButton(this, "MAIN_SCREEN_SHOW_COMMS_BUTTON", tr("MainScreenControls", "Show comms"), [this]()
         {
             if (my_spaceship)
             {
@@ -139,7 +139,7 @@ GuiMainScreenControls::GuiMainScreenControls(GuiContainer* owner)
         }));
         show_comms_button = buttons.back();
 
-        buttons.push_back(new GuiButton(this, "MAIN_SCREEN_HIDE_COMMS_BUTTON", "Hide comms", [this]()
+        buttons.push_back(new GuiButton(this, "MAIN_SCREEN_HIDE_COMMS_BUTTON", tr("MainScreenControls", "Hide comms"), [this]()
         {
             if (my_spaceship)
             {