+++ src/screenComponents/scanTargetButton.cpp
@@ -9,7 +9,7 @@
 GuiScanTargetButton::GuiScanTargetButton(GuiContainer* owner, string id, TargetsContainer* targets)
 : GuiElement(owner, id), targets(targets)
 {
-    button = new GuiButton(this, id + "_BUTTON", "Scan", [this]() {
+    button = new GuiButton(this, id + "_BUTTON", tr("Scan"), [this]() {
         if (my_spaceship && this->targets && this->targets->get())
             my_spaceship->commandScan(this->targets->get());
     });